Add experimental time freezing support.
Datetime can now be used as a context manager to freeze time.
The freeze class method can also be used as decorator to mimic the behaviour
of e.g. the freezegun package.

There's also some experiments probing the _potential_ change in behaviour for comparisons such as `==`. Not so sure that those will make the cut yet though.